requirements.txt
openai==1.60.2
langchain==0.3.15
langchain-community==0.3.15
pillow==10.4.0
tiktoken==0.8.0
matplotlib==3.10.0
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
requests==2.32.3
python-dotenv==1.0.1
transformers==4.48.3
pandas==2.2.3
pinecone==6.0.1
sentence_transformers
google-generativeai==0.8.4
langchain-community==0.3.17